All mappers have been transferred to INSTANCE

+++ src/main/java/com/smalko/scoreboard/controller/MatchesController.java
@@ -20,8 +20,8 @@ public class MatchesController {
     private MatchesController() {
     }
 
-    public int countMatch(){
-        int count;
+    public long countMatch(){
+        long count;
         try (SessionFactory sessionFactory = HibernateUtil.sessionFactory()) {
             var session = (Session) Proxy.newProxyInstance(SessionFactory.class.getClassLoader(), new Class[]{Session.class},
                     (proxy, method, args1) -> method.invoke(sessionFactory.getCurrentSession(), args1));
